puzzles/python_requirements.txt
appdirs==1.4.3
appnope==0.1.0
click==6.7
decorator==4.0.11
EasyProcess==0.2.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-dotenv==0.6.4
selenium==3.4.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
